requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
exchange_calendars==4.11
fonttools==4.59.1
fredapi==0.5.2
frozendict==2.4.6
idna==3.10
kiwisolver==1.4.9
korean-lunar-calendar==0.3.1
matplotlib==3.10.5
multitasking==0.0.12
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pandas_market_calendars==5.1.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.32.0
pycparser==2.22
pyluach==2.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scipy==1.16.1
six==1.17.0
soupsieve==2.7
toolz==1.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
